// src/_workspace/types/bom/Bom.ts
//* Convert to type

export interface BomI {
  BOM_CODE: string
  BOM_ID: number
  BOM_NAME: string
  FLOW_CODE: string
  FLOW_ID: number
  FLOW_NAME: string
  INUSE: number
  INUSE_RAW_DATA: number
  MODIFIED_DATE: string
  No: number
  PRODUCTION_PURPOSE_ALPHABET: string
  PRODUCTION_PURPOSE_ID: number
  PRODUCTION_PURPOSE_NAME: string
  PRODUCT_CATEGORY_ID: number
  PRODUCT_CATEGORY_NAME: string
  PRODUCT_MAIN_ALPHABET: string
  PRODUCT_MAIN_ID: number
  PRODUCT_MAIN_NAME: string
  REVISION: string
  TOTAL_COUNT_PROCESS: number
  UPDATE_BY: string
  PRODUCT_TYPE_CODE: string
  PRODUCT_TYPE_NAME: string
  IS_DRAFT: number
}